requirements.txt
altair==4.1.0
altair-saver==0.5.0
beautifulsoup4==4.9.3
dateparser==1.0.0
lxml==4.6.3
mdutils==1.3.0
openpyxl==3.0.7
pandas==1.2.3
python-pushover==0.4
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0